# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DopaFlow API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v2/journal/entries": {
      "get": {
        "tags": [
          "journal"
        ],
        "summary": "List Entries",
        "description": "List journal entries with optional tag or full-text search filter.",
        "operationId": "list_entries_api_v2_journal_entries_get",
        "parameters": [
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JournalEntryRead"
                  },
                  "title": "Response List Entries Api V2 Journal Entries Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "journal"
        ],
        "summary": "Save Entry",
        "description": "Create or update a journal entry for the given date.",
        "operationId": "save_entry_api_v2_journal_entries_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JournalEntryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalEntryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/journal/entries/{identifier}": {
      "get": {
        "tags": [
          "journal"
        ],
        "summary": "Get Entry",
        "description": "Fetch a single journal entry by its ID or date string (YYYY-MM-DD).",
        "operationId": "get_entry_api_v2_journal_entries__identifier__get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalEntryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "journal"
        ],
        "summary": "Delete Entry",
        "description": "Soft-delete a journal entry by ID or date.",
        "operationId": "delete_entry_api_v2_journal_entries__identifier__delete",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/journal/backup/trigger": {
      "post": {
        "tags": [
          "journal"
        ],
        "summary": "Trigger Backup",
        "description": "Write the specified date's journal entry to the local markdown backup.",
        "operationId": "trigger_backup_api_v2_journal_backup_trigger_post",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "YYYY-MM-DD \u2014 defaults to today",
              "title": "Date"
            },
            "description": "YYYY-MM-DD \u2014 defaults to today"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalBackupTriggerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/journal/{identifier}/backlinks": {
      "get": {
        "tags": [
          "journal"
        ],
        "summary": "Backlinks",
        "description": "Return entry IDs that contain a wikilink to the given date slug.",
        "operationId": "backlinks_api_v2_journal__identifier__backlinks_get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Backlinks Api V2 Journal  Identifier  Backlinks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JournalBackupTriggerResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "backed_up_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backed Up Date"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "JournalBackupTriggerResponse",
        "description": "Confirmation that a manual backup was triggered."
      },
      "JournalDeleteResponse": {
        "properties": {
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          },
          "identifier": {
            "type": "string",
            "title": "Identifier"
          }
        },
        "type": "object",
        "required": [
          "deleted",
          "identifier"
        ],
        "title": "JournalDeleteResponse",
        "description": "Response after soft-deleting a journal entry."
      },
      "JournalEntryCreate": {
        "properties": {
          "markdown_body": {
            "type": "string",
            "title": "Markdown Body"
          },
          "emoji": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emoji"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "markdown_body",
          "date"
        ],
        "title": "JournalEntryCreate",
        "description": "Payload for creating or updating a journal entry."
      },
      "JournalEntryRead": {
        "properties": {
          "markdown_body": {
            "type": "string",
            "title": "Markdown Body"
          },
          "emoji": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emoji"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          },
          "locked": {
            "type": "boolean",
            "title": "Locked",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "markdown_body",
          "date",
          "id"
        ],
        "title": "JournalEntryRead",
        "description": "Serialized journal entry returned from the API."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}